Refactor
1. Added a TEST flag. If testing, no uploads / moving of files are done, just logs printed to debug.

2. Removed Settings object. Details of individual decks are contained within their own classes instead of within Settings object.

+++ models/FileMover.cs
@@ -11,26 +11,28 @@ namespace YouTubeUploadBot.models
     {
         static Logger logger = LogManager.GetCurrentClassLogger();
 
-        // TODO: GetDestinationFolder is called twice??
-
-        public static void MoveUploadedFileNew(Settings settings, string pathToUploadedFile, Deck myDeck)
+        public static void MoveUploadedFile(string pathToUploadedFile, Deck myDeck, Boolean testMove)
         {
             try
             {
                 FileInfo fileInfo = new FileInfo(pathToUploadedFile);
 
                 string destinationFolder = myDeck.uploadedFolder;
+                logger.Trace($"destinationFolder: {destinationFolder}");
 
                 string srcPath = fileInfo.FullName;
-                logger.Trace($"srcPath: {srcPath}");
+                logger.Trace($"srcPath: {srcPath} {Environment.NewLine}");
 
-                string dstPath = CreateDestinationPathNew(settings, myDeck, srcPath);
+                string dstPath = CreateDestinationPath(myDeck, srcPath, testMove);
                 logger.Trace($"dstPath: {dstPath}");
 
-                File.Delete(srcPath);
-                Directory.Delete(Path.GetDirectoryName(srcPath));
+                if (!testMove)
+                {
+                    File.Delete(srcPath);
+                    Directory.Delete(Path.GetDirectoryName(srcPath));
 
-                logger.Info($"File has been moved from {srcPath} to {dstPath}. {Environment.NewLine}");
+                    logger.Info($"File has been moved from {srcPath} to {dstPath}. {Environment.NewLine}");
+                }
             }
             catch (Exception e)
             {
@@ -38,7 +40,7 @@ namespace YouTubeUploadBot.models
             }
         }
 
-        public static string CreateDestinationPathNew(Settings settings, Deck myDeck, string filePath)
+        public static string CreateDestinationPath(Deck myDeck, string filePath, Boolean testMove)
         {
             logger.Trace($"deckName:            {myDeck.deckName}");
             logger.Trace($"filePath:            {filePath}");
@@ -65,99 +67,22 @@ namespace YouTubeUploadBot.models
 
                 logger.Trace($"Trying {nextAvailableDirectory}...");
             }
-            Directory.CreateDirectory(nextAvailableDirectory);
-            logger.Info($"Created directory: {nextAvailableDirectory}");
+
+            if (!testMove)
+            {
+                Directory.CreateDirectory(nextAvailableDirectory);
+                logger.Info($"Created directory: {nextAvailableDirectory}");
+            }
+            else
+            {
+                logger.Info($"Directory would be created at: {nextAvailableDirectory}");
+            }
 
             nextAvailableDirectory = Path.Combine(nextAvailableDirectory, fileName);
-            logger.Trace($"Returning path: {nextAvailableDirectory}");
+            logger.Trace($"Returning path: {nextAvailableDirectory} {Environment.NewLine}");
             // G:\YouTube\MTG\footage\Grixis Truths\Uploaded\Jund Midrange 02\Jund Midrange.mp4
 
             return nextAvailableDirectory;
-        }
-
-        public static void MoveUploadedFile(Settings settings, string pathToUploadedFile, string myDeck)
-        {
-            try
-            {
-                FileInfo fileInfo = new FileInfo(pathToUploadedFile);
-
-                string destinationFolder = GetDestinationFolder(settings, myDeck);
-                logger.Trace($"destinationFolder: {destinationFolder}");
-
-                string srcPath = fileInfo.FullName;
-                logger.Trace($"srcPath: {srcPath}");
-
-                string dstPath = CreateDestinationPath(settings, myDeck, srcPath);
-                logger.Trace($"dstPath: {dstPath}");
-
-                File.Copy(srcPath, dstPath, false);
-
-                File.Delete(srcPath);
-                Directory.Delete(Path.GetDirectoryName(srcPath));
-
-                logger.Info($"File has been moved from {srcPath} to {dstPath}. {Environment.NewLine}");
-            }
-            catch (Exception e)
-            {
-                logger.Error(e);
-            }
-        }
-
-        // receives: "G:\YouTube\MTG\footage\Upload From Here\06 Jeskai Truths vs Izzet Wizards\Izzet Wizards.mp4"
-        // returns:  "G:\YouTube\MTG\footage\Jeskai Truths\Uploaded\Izzet Wizards 01"
-        public static string CreateDestinationPath(Settings settings, string myDeck, string filePath)
-        {
-            logger.Trace($"myDeck:      {myDeck}");
-            logger.Trace($"filePath:    {filePath}");
-            
-            string destinationFolder = GetDestinationFolder(settings, myDeck);
-            logger.Trace($"destinationFolder: {destinationFolder}");
-
-            string fileName = Path.GetFileName(filePath);
-            logger.Trace($"fileName: {fileName}");
-
-            string fileNameWithoutExtension = Path.GetFileNameWithoutExtension(filePath);
-            logger.Trace($"fileNameWithoutExtension: {fileNameWithoutExtension}");
-
-            int start = 1;
-
-            string nextAvailableDirectory = Path.Combine(destinationFolder, fileNameWithoutExtension + " " + start.ToString("D2"));
-            logger.Info($"Checking availability of: {nextAvailableDirectory}...");
-            // G:\YouTube\MTG\footage\Grixis Truths\Uploaded\Jund Midrange 01
-
-            while (Directory.Exists(nextAvailableDirectory))
-            {
-                logger.Trace($"Folder {nextAvailableDirectory} already exists!");
-                start++;
-                nextAvailableDirectory = Path.Combine(destinationFolder, fileNameWithoutExtension + " " + start.ToString("D2"));
-                // G:\YouTube\MTG\footage\Grixis Truths\Uploaded\Jund Midrange 02
-
-                logger.Trace($"Trying {nextAvailableDirectory}...");
-            }
-            Directory.CreateDirectory(nextAvailableDirectory);
-            logger.Info($"Created directory: {nextAvailableDirectory}");
-
-            nextAvailableDirectory = Path.Combine(nextAvailableDirectory, fileName);
-            logger.Trace($"Returning path: {nextAvailableDirectory}");
-            // G:\YouTube\MTG\footage\Grixis Truths\Uploaded\Jund Midrange 02\Jund Midrange.mp4
-
-            return nextAvailableDirectory;
-        }
-
-        public static string GetDestinationFolder(Settings settings, string myDeck)
-        {
-            switch(myDeck)
-            {
-                case "Jeskai Truths":
-                    return settings.jeskaiTruths.uploadedFolder;
-                case "Grixis Truths":
-                    return settings.grixisTruths.uploadedFolder;
-                case "Bant Yorion":
-                    return settings.bantYorion.uploadedFolder;
-                default:
-                    logger.Warn($"{myDeck} has no settings configured!");
-                    return String.Empty;
-            }
         }
     }
 }
